Test read-lock->write-lock->read-lock behavior
When one process read locks a range of a file, a second process bloks
while trying to get a write lock on the same range, and a third
processes then attempts to get another read lock on the same range will
be allowed. This behavior suffers the risk to starve processes with
pending write locks.